chore(lint): clippy-clean the workspace, with guest-crate lints in the manifests
CI runs `cargo clippy --all-targets -- -D warnings` on the pinned
nightly-2026-07-17, so every warning is a red gate.  271 distinct sites ->
68; nothing here changes behaviour.

Guest-reachable crates (`zkvm/lib`, `zkvm/entrypoint`, `primitives`) get
their lint levels in Cargo.toml rather than in-source `#[allow]`:
reth-processor patches `zkm-zkvm`/`zkm-lib` to path deps on this tree, so
editing those sources rebuilds the production guest ELF -- and moves the
program vk -- for a style fix.  A manifest lint level reaches rustc as a
flag and cannot perturb guest codegen.  `unused_features` there is a
host-build artifact: the `asm!` blocks consuming
`feature(asm_experimental_arch)` are `cfg(target_os = "zkvm")`.

`type_complexity` is allowed per crate with its reason: the jagged/BaseFold
generics (Challenger x Mmcs x Dft over two rings) are wide because the ring
a function is instantiated for should be readable off its signature, which
is what the soundness review needs.

Two defects surfaced, both doc-correctness rather than style:

  * `prove_shard_zerocheck` was undocumented while its pipeline doc AND its
    `#[allow(too_many_arguments)]` sat above a ONE-argument helper inserted
    between them.  Both moved back onto the function.
  * a stale "Capabilities query" block was silently prepended to
    `JIT_MIN_INSTR_COUNT`'s rustdoc; `first_unsupported_opcode` already
    carries its own, accurate doc.  Removed.

Also: per-item docs for the MIPS lowering API and register file (both files
documented only the first member of each group, so the rest inherited
nothing) -- `srlv` = `rd = rt >> (rs mod 32)` zero-fill, `bltz` = `rs <_s 0`
under a shared target rule, `lwl`/`lwr` by which `4 - (a mod 4)` bytes each
merges.  `Lo`/`Hi` state the invariant they exist for (the discriminant must
equal the executor's `Register::LO = 32`) instead of the incident that found
it.  Guarded divisions become `checked_div(..).unwrap_or(..)`, and
`Default::default()`-then-assign becomes struct literals.  Eighteen
content-free label comments removed.

Verified on ant-5090-2: `cargo check -r --workspace --all-targets` clean;
zkm-pcs (jagged), zkm-recursion-core, zkm-core-executor suites pass.
zkm-verifier's 4 vk fixtures fail on the STALE vk_map (pending the
campaign), unrelated to this change: every AIR-file edit here is inert
(unused-binding renames, `chunks_exact(4)` -> `as_chunks::<4>().0` over the
same full chunks, `iter_mut()` -> `values_mut()`, clone removal on `Copy`
field types, one redundant `.into()`).

+++ crates/core/jit/src/driver.rs
@@ -297,7 +297,7 @@ pub(crate) fn lower_one<T: MipsTranspiler>(
     // lowering expects an *absolute* target PC (per the existing
     // comment on beq/bne in instruction_impl.rs), so we resolve it
     // here once and pass the absolute value down.
-    let branch_target_pc = (current_pc.wrapping_add(4).wrapping_add(ins.op_c as u32)) as i32;
+    let branch_target_pc = (current_pc.wrapping_add(4).wrapping_add(ins.op_c)) as i32;
 
     match op {
         // ALU
@@ -340,7 +340,7 @@ pub(crate) fn lower_one<T: MipsTranspiler>(
         JitOpcode::Srl => {
             if ins.imm_b && ins.imm_c {
                 let shamt = (ins.op_c & 0x1f) as u8;
-                let val = (ins.op_b as u32).wrapping_shr(shamt as u32);
+                let val = ins.op_b.wrapping_shr(shamt as u32);
                 t.add(rd, MipsOperand::Imm(val as i64), MipsOperand::Imm(0));
             } else if ins.imm_c {
                 t.srl(rd, rs, (ins.op_c & 0x1f) as u8);
@@ -362,7 +362,7 @@ pub(crate) fn lower_one<T: MipsTranspiler>(
         JitOpcode::Ror => {
             if ins.imm_b && ins.imm_c {
                 let shamt = (ins.op_c & 0x1f) as u8;
-                let val = (ins.op_b as u32).rotate_right(shamt as u32);
+                let val = ins.op_b.rotate_right(shamt as u32);
                 t.add(rd, MipsOperand::Imm(val as i64), MipsOperand::Imm(0));
             } else if ins.imm_c {
                 t.ror(rd, rs, (ins.op_c & 0x1f) as u8);
@@ -372,7 +372,7 @@ pub(crate) fn lower_one<T: MipsTranspiler>(
         }
         JitOpcode::Clz => {
             if ins.imm_b {
-                let val = (ins.op_b as u32).leading_zeros();
+                let val = ins.op_b.leading_zeros();
                 t.add(rd, MipsOperand::Imm(val as i64), MipsOperand::Imm(0));
             } else {
                 t.clz(rd, rs);
@@ -380,7 +380,7 @@ pub(crate) fn lower_one<T: MipsTranspiler>(
         }
         JitOpcode::Clo => {
             if ins.imm_b {
-                let val = (ins.op_b as u32).leading_ones();
+                let val = ins.op_b.leading_ones();
                 t.add(rd, MipsOperand::Imm(val as i64), MipsOperand::Imm(0));
             } else {
                 t.clo(rd, rs);
